chore(FTDA): File structure are changed according to conventions. chore(FTDA): Remove unnecessary lines. chore(FTDA): Imports are done according to conventions
