+++ src/App.tsx
@@ -1,33 +1,16 @@
-import "./index.css";
 import "@mantine/core/styles.css";
-import { Button, Container, MantineProvider } from "@mantine/core";
-import { useState } from "react";
-import ToDoForm from "./components/ToDoForm";
+import { MantineProvider } from "@mantine/core";
 import '@mantine/dates/styles.css';
+
+import "./index.css";
 import { TasksProvider } from "./contexts/TasksContext";
-import "./index.css"
+import HomePageContainer from "./Pages/HomePage/Containers/HomePageContainer";
 
 function App() {
-  const handleToDoForm = () => {
-    setOpened(true);
-  };
-
-  const [opened, setOpened] = useState(false);
   return (
     <TasksProvider>
       <MantineProvider defaultColorScheme="dark">
-        <Container>
-          <h1 className="title">Fancy To-DO With TypeScript</h1>
-          <Button
-            fullWidth
-            variant="gradient"
-            gradient={{ from: "blue", to: "cyan", deg: 90 }}
-            onClick={handleToDoForm}
-          >
-            Add Task
-          </Button>
-          <ToDoForm opened={opened} setOpened={setOpened}/>
-        </Container>
+        <HomePageContainer />
       </MantineProvider>
     </TasksProvider>
   );
